+++ packages/firebase_storage/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.3.1
+
+* Added support for StorageReference functions: `getParent`, `getRoot`, `getStorage`, `getName`, `getPath`, `getBucket`.
+
 ## 0.3.0
 
 * **Breaking change**. Changed StorageUploadTask to abstract, removed the 'file' field, and made 'path' and 'metadata'